Avoid build warnings on technical issues (corrected) + set version to "1.1.0"

+++ WorkersRightsComputer/UsedVacations.cpp
@@ -281,7 +281,7 @@ void CUsedVacations::Log()
 }
 void CUsedVacations::WriteToLetter(CHtmlWriter& html)
 {
-	int nVacations = mVacationsUsed.GetSize();
+	int nVacations = (int)mVacationsUsed.GetSize();
 	if (nVacations < 1)
 		return;
